LayoutTests/platform/blackberry/fast/dom/SelectorAPI/caseTagX-expected.txt

CONSOLE MESSAGE: line 18: SyntaxError: Parse error
FAIL successfullyParsed should be true. Threw exception ReferenceError: Can't find variable: successfullyParsed

TEST COMPLETE
